requirements.txt
aiogram==3.4.1
python-dotenv==1.0.1

psycopg2-binary==2.9.9
django==4.2.13
djangorestframework==3.15.1